stok in edit page updated

+++ pharmacist/ajax/stockin.delete.ajax.php
@@ -4,14 +4,20 @@ ini_set('display_startup_errors', '1');
 error_reporting(E_ALL);
 
 require_once '../../php_control/stockIn.class.php';
+require_once '../../php_control/stockInDetails.class.php';
 
 $Stockin = new StockIn();
+$stockInDetails = new StockInDetails();
 
     $stockId  = $_POST['Currentid'];
 
-    $stockDelete = $Stockin->deleteStock($stockId);
+    $stockInDetailsDelete = $stockInDetails->stockInDeletebyId($stockId);
 
-    if($stockDelete){
+    if($stockInDetailsDelete == true){
+        $stockDelete = $Stockin->deleteStock($stockId);
+    }
+
+    if($stockDelete == true){
         echo 1;
     }else{
         echo 0;
